remove support for async fn
