+++ packages/core/src/nodes/dynamic_nodes.rs
@@ -1,0 +1,1 @@
+use crate::VNode;
